+++ Assets/Scripts/Items/ShieldBoost.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ShieldBoost : Effect
 {
-    public ShieldBoost(int _duration, float _dshield, int _priority, int _key)
+    public ShieldBoost(float _duration, float _dshield, int _priority, int _key)
     {
         duration = _duration;
         dshield = _dshield;
